requirements.txt
# API
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1

# Config & HTTP client (API + Streamlit)
python-dotenv==1.1.0
httpx==0.28.1
requests==2.32.3
tenacity==9.1.2

# ML — TrustScoreMLP / training (`app/ml`) + HF pipelines (`app/services/explanation_insights.py`)
torch>=2.2.0
transformers>=4.45.0,<5

# Streamlit demo (`streamlit_app.py`)
streamlit>=1.39.0
plotly>=5.24.0
